+++ R/data_plot.R
@@ -71,6 +71,7 @@
 #'   \item \code{bayestestR::hdi()}
 #'   \item \code{bayestestR::map_estimate()}
 #'   \item \code{bayestestR::p_direction()}
+#'   \item \code{bayestestR::p_significance()()}
 #'   \item \code{bayestestR::point_estimate()}
 #'   \item \code{bayestestR::rope()}
 #'   \item \code{estimate::estimate_contrasts()}